+++ src/main/handlers/pathsHandler.ts
@@ -10,7 +10,6 @@ ipcMain.handle("get-library-path", async () => {
 
 ipcMain.handle("set-library-path", async (_event, libraryPath: string | null) => {
   store.set("libraryPath", libraryPath);
-  console.log("Library path set to:", libraryPath);
 });
 
 ipcMain.handle("get-target-path", async () => {
@@ -19,34 +18,31 @@ ipcMain.handle("get-target-path", async () => {
 
 ipcMain.handle("set-target-path", async (_event, targetPath: string | null) => {
   store.set("targetPath", targetPath);
-  console.log("Target path set to:", targetPath);
 });
 
 ipcMain.handle("open-mod-folder", async (_event, modName?: string) => {
   const libraryPath = store.get("libraryPath", null) as string | null;
-  if (!libraryPath) return;
-  const fullPath = modName ? path.resolve(libraryPath, modName) : libraryPath;
+  if (!libraryPath || !(await fs.pathExists(libraryPath))) return;
+  const fullPath = modName ? path.join(libraryPath, modName) : libraryPath;
   await shell.openPath(fullPath);
 });
 
 ipcMain.handle("open-target-folder", async () => {
   const targetPath = store.get("targetPath", null) as string | null;
-  if (!targetPath) return;
+  if (!targetPath || !(await fs.pathExists(targetPath))) return;
   await shell.openPath(targetPath);
 });
 
 ipcMain.handle("clear-target-path", async () => {
   // remove all symlinks ONLY
   const targetPath = store.get("targetPath", null) as string | null;
-  if (!targetPath) return;
+  if (!targetPath || !(await fs.pathExists(targetPath))) return;
   try {
-    const files = await fs.readdir(targetPath);
+    const files = await fs.readdir(targetPath, { withFileTypes: true });
     for (const file of files) {
-      const filePath = path.join(targetPath, file);
-      const stats = await fs.lstat(filePath);
-      if (stats.isSymbolicLink()) {
-        await fs.unlink(filePath);
-      }
+      const filePath = path.join(targetPath, file.name);
+      if (!file.isSymbolicLink()) continue;
+      await fs.unlink(filePath);
     }
   } catch (error) {
     console.error("Error clearing target path:", error);
